+++ src/docker/composerunner/image_group.go
@@ -1,4 +1,4 @@
-package runner
+package composerunner
 
 // ImageGroup is a group of images to run at once
 type ImageGroup struct {
