+++ src/docker/composerunner/run.go
@@ -1,4 +1,4 @@
-package runner
+package composerunner
 
 import (
 	"fmt"
