+++ src/docker/composerunner/run_image_group.go
@@ -1,4 +1,4 @@
-package runner
+package composerunner
 
 import (
 	"fmt"
