+++ src/docker/composerunner/shutdown.go
@@ -1,4 +1,4 @@
-package runner
+package composerunner
 
 import (
 	"fmt"
